+++ udata_front/theme/gouvfr/assets/js/components/dataset/card-lg.vue
@@ -64,65 +64,7 @@
             >
               {{$t('Metadata quality:')}}
               <template #toggletip>
-                <h5 class="fr-text--sm fr-my-0">{{$t("Metadata quality:")}}</h5>
-                <QualityItem
-                  :passed="dataset.quality.dataset_description_quality"
-                  :messagePassed='$t("Data description filled")'
-                  :messageFailed='$t("Data description empty")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.resources_documentation"
-                  :messagePassed='$t("Files documented")'
-                  :messageFailed='$t("Files documentation missing")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.license"
-                  :messagePassed='$t("License filled")'
-                  :messageFailed='$t("No license set")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.update_frequency && !!dataset.quality.update_fulfilled_in_time"
-                  :messagePassed='$t("Update frequency followed")'
-                  :messageFailed='dataset.quality.update_frequency ? $t("Update frequency not followed") : $t("Update frequency not set")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.has_open_format"
-                  :messagePassed='$t("File formats are open")'
-                  :messageFailed='$t("File formats are closed")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.temporal_coverage"
-                  :messagePassed='$t("Temporal coverage filled")'
-                  :messageFailed='$t("Temporal coverage not set")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.spatial"
-                  :messagePassed='$t("Spatial coverage filled")'
-                  :messageFailed='$t("Spatial coverage not set")'
-                  class="fr-my-1w"
-                />
-                <QualityItem
-                  :passed="dataset.quality.all_resources_available"
-                  :messagePassed='$t("All files are available")'
-                  :messageFailed='$t("Some files are unavailable")'
-                  class="fr-my-1w"
-                />
-                <div class="fr-grid-row fr-grid-row--right not-enlarged">
-                  <a
-                    :href="guides_quality_url"
-                    target="_blank"
-                    rel="noopener"
-                    :title="$t('Learn more about this indicator - opens a new window')"
-                  >
-                    {{$t("Learn more about this indicator")}}
-                  </a>
-                </div>
+                <QualityScoreTooltipContent :dataset="dataset"/>
               </template>
             </Toggletip>
             <p class="fr-m-0 fr-mr-1v" :id="id">
@@ -170,6 +112,7 @@ import Placeholder from "../utils/placeholder.vue";
 import QualityScore from "./quality-score.vue";
 import Toggletip from "../utils/Toggletip/Toggletip.vue";
 import QualityItem from "./quality-item.vue";
+import QualityScoreTooltipContent from "./QualityScore/QualityScoreTooltipContent/QualityScoreTooltipContent.vue";
 import { excerpt } from "../../helpers";
 import { guides_quality_url, quality_metadata_backend_ignore } from "../../config";
 import type { Dataset } from "../../types";
@@ -180,6 +123,7 @@ export default defineComponent({
     OrganizationNameWithCertificate,
     Placeholder,
     QualityScore,
+    QualityScoreTooltipContent,
     Toggletip,
     QualityItem,
 },